refactor: repository에 대한 인터페이스 생성
